requirements.txt
Appium-Python-Client==2.6.1
pytest-bdd==6.1.1
pytest-html==3.2.0
selenium==4.6.0


